// 2020/website/src/map/interfaces.d.ts
export type APObj = {
  [key: string]: any;
  latitude: number;
  longitude: number;
  altitude?: number;
  radius?: number;
  units?: string;
  accuracy?: number;
};
export type LngObj = { lat: number; lng: number; alt?: number; };
export type LatLng = LeafletLatLng | [number, number] | APObj | LngObj;

export type LatLngBounds = LeafletLatLngBounds | Array<LatLng>

export type ControlPosition =
  | 'topleft'
  | 'topright'
  | 'bottomleft'
  | 'bottomright'

export type Point = [number, number] | LeafletPoint

export interface Viewport {
  center?: [number, number],
  zoom?: number,
}
// ...
export type CtrlPosition = boolean | ControlPosition;
export type ZoomOption = boolean | 'center';
export type Props = {
  [key: string]: any;
  // Leaflet options
  preferCanvas?: boolean;
  attributionControl?: CtrlPosition;
  zoomControl?: CtrlPosition;
  zoomSnap?: number;
  zoomDelta?: number;
  closePopupOnClick?: boolean;
  trackResize?: boolean;
  boxZoom?: boolean;
  doubleClickZoom?: ZoomOption;
  dragging?: boolean;
  crs?: CRS;
  center?: LatLng;
  zoom?: number;
  minZoom?: number;
  maxZoom?: number;
  maxBounds?: any;//LatLngBounds;
  renderer?: Renderer;
  zoomAnimation?: boolean;
  zoomAnimationThreshold?: number;
  fadeAnimation?: boolean;
  markerZoomAnimation?: boolean;
  transform3DLimit?: number;
  inertia?: boolean;
  inertiaDeceleration?: number;
  inertiaMaxSpeed?: number;
  easeLinearity?: number;
  worldCopyJump?: boolean;
  maxBoundsViscosity?: number;
  keyboard?: boolean;
  keyboardPanDelta?: number;
  scrollWheelZoom?: ZoomOption;
  wheelDebounceTime?: number;
  wheelPxPerZoomLevel?: number;
  tap?: boolean;
  tapTolerance?: number;
  touchZoom?: ZoomOption;
  bounceAtZoomLimits?: boolean;

  // Additional options
  animate?: boolean;
  duration?: number;
  noMoveStart?: boolean;
  bounds?: LatLngBounds;
  boundsOptions?: {
    paddingTopLeft?: Point;
    paddingBottomRight?: Point;
    padding?: Point;
    maxZoom?: number;
  };
  children?: Node;
  className?: string;
  id?: string;
  style?: Object;
  useFlyTo?: boolean;
  viewport?: Viewport;
  whenReady?: () => void;
}